% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced.hazard.ratio.R
\name{balanced.hazard.ratio}
\alias{balanced.hazard.ratio}
\title{Function to estimate the balanced hazard ratio through Cox regression}
\usage{
balanced.hazard.ratio(x, surv.time, surv.event, alpha = 0.05,
method.test = c("logrank", "likelihood.ratio", "wald"),
ties = c("efron", "breslow", "exact"), weights, strat, ...)
}
\arguments{
\item{x}{a vector of risk group predictions.}

\item{surv.time}{a vector of event times.}

\item{surv.event}{a vector of event occurrence indicators.}

\item{alpha}{alpha level to compute confidence interval.}

\item{method.test}{...}

\item{ties}{...}

\item{weights}{...}

\item{strat}{...}

\item{...}{Additional parmaeters to the hazard.ratio and coxph functions.}
}
\value{
A list with the items:
\itemize{
\item balanced.hazard.ratio: balanced hazard ratio estimate.
\item coef: coefficient (beta) estimated in the cox regression model.
\item se: standard error of the coefficient (beta) estimate.
\item lower: lower bound for the confidence interval.
\item upper: upper bound for the confidence interval.
\item p.value: p-value computed using the score (logrank) test whether the
balanced hazard ratio is different from 1.
\item n: number of samples used for the estimation.
\item coxm: coxph.object fitted on the survival data and x.
\item data: list of data used to compute the balanced hazard ratio (x,
surv.time and surv.event).
}
}
\description{
Function to compute the balanced hazard ratio for a risk group prediction.
The balanced hazard ratio is computed using the Cox model.
}
\examples{
set.seed(12345)
age <- rnorm(100, 50, 10)
stime <- rexp(100)
cens   <- runif(100,.5,2)
sevent  <- as.numeric(stime <= cens)
stime <- pmin(stime, cens)
strat <- sample(1:3, 100, replace=TRUE)
weight <- runif(100, min=0, max=1)
balanced.hazard.ratio(x=age, surv.time=stime, surv.event=sevent,
  weights=weight, strat=strat)

}
\references{
Branders, S. and Dupont, P. (2015) "A balanced hazard ratio for risk group
evaluation from survival data", Statistics in Medicine, 34(17),
pages 2528–2543.
}
\seealso{
\link{hazard.ratio}, \link[survival:coxph]{survival::coxph}, \link[survival:coxph.object]{survival::coxph.object}
}
