% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhr.comp.R
\name{bhr.comp}
\alias{bhr.comp}
\title{Function to statistically compare two balanced hazard ratios}
\usage{
bhr.comp(bhr1, bhr2)
}
\arguments{
\item{bhr1}{first balanced hazard ratio.}

\item{bhr2}{second balanced hazard ratio.}
}
\value{
A list with items:
\itemize{
\item p.value:	p-value from the Student t test for the comparison beta1 >
beta2 (equivalently bhr1 > bhr2)
\item bhr1: value of the first balanced hazard ratio
\item bhr2: value of the second balanced hazard ratio
}
}
\description{
This function compares two balanced hazard ratios from their betas and
standard errors as computed by a Cox model for instance. The statistical
test is a Student t test for dependent samples. The two balanced hazard
ratios must be computed from the same survival data.
}
\details{
The two balanced hazard ratios must be computed from the same samples
(and corresponding survival data). The function uses a Student t test for
dependent samples.
}
\examples{
set.seed(12345)
age <- as.numeric(rnorm(100, 50, 10) >= 50)
size <- as.numeric(rexp(100,1) > 1)
stime <- rexp(100)
cens <- runif(100,.5,2)
sevent <- as.numeric(stime <= cens)
stime <- pmin(stime, cens)
bhr1 <- balanced.hazard.ratio(x=age, surv.time=stime, surv.event=sevent)
bhr2 <- balanced.hazard.ratio(x=size, surv.time=stime, surv.event=sevent)
bhr.comp(bhr1=bhr1, bhr2=bhr2)

}
\references{
Student. (1908) "The Probable Error of a Mean", Biometrika, 6, 1, pages 1–25.
Haibe-Kains, B. and Desmedt, C. and Sotiriou, C. and Bontempi, G. (2008)
"A comparative study of survival models for breast cancer prognostication
based on microarray data: does a single gene beat them all?",
Bioinformatics, 24, 19, pages 2200–2208.
Branders, S. and Dupont, P. (2015) "A balanced hazard ratio for risk group
evaluation from survival data", Statistics in Medicine, 34(17),
pages 2528–2543.
}
\seealso{
\link{balanced.hazard.ratio}
\link[survival:coxph]{survival::coxph}
\link[stats:t.test]{stats::t.test}
}
