% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cindex.comp.R
\name{cindex.comp}
\alias{cindex.comp}
\title{Function to compare two concordance indices}
\usage{
cindex.comp(cindex1, cindex2)
}
\arguments{
\item{cindex1}{first concordance index as returned by the concordance.index function.}

\item{cindex2}{second concordance index as returned by the concordance.index function.}
}
\value{
A list with items:
\itemize{
\item p.value: p-value from the Student t test for the comparison cindex1 >
\item cindex1: value of the first concordance index.
\item cindex2: value of the second concordance index.
}
}
\description{
This function compares two concordance indices computed from the same data by
using the function concordance.index. The statistical test is a Student
t-test for dependent samples.
}
\examples{
set.seed(12345)
age <- rnorm(100, 50, 10)
size <- rexp(100,1)
stime <- rexp(100)
cens <- runif(100,.5,2)
sevent <- as.numeric(stime <= cens)
stime <- pmin(stime, cens)
c1 <- concordance.index(x=age, surv.time=stime, surv.event=sevent,
  method="noether")
c2 <- concordance.index(x=size, surv.time=stime, surv.event=sevent,
  method="noether")
cindex.comp(c1, c2)

}
\references{
Haibe-Kains, B. and Desmedt, C. and Sotiriou, C. and Bontempi, G. (2008)
"A comparative study of survival models for breast cancer prognostication
based on microarray data: does a single gene beat them all?",
Bioinformatics, 24(19), pages 2200–2208.
}
\seealso{
\link{concordance.index}
}
