% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.est.R
\name{combine.est}
\alias{combine.est}
\title{Function to combine estimates}
\usage{
combine.est(x, x.se, hetero = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{vector of estimates}

\item{x.se}{vector of standard errors of the corresponding estimates}

\item{hetero}{\code{TRUE} is the heterogeneity should be taken into account
(random effect model), \code{FALSE} otherwise (fixed effect model)}

\item{na.rm}{TRUE if the missing values should be removed from the data,
\code{FALSE} otherwise.}
}
\value{
A list with items:
\itemize{
\item estimate: meta-estimate
\item se: standard error of the meta-estimate
}
}
\description{
The function combines several estimators using meta-analytical formula to
compute a meta-estimate.
}
\examples{
set.seed(12345)
x1 <- rnorm(100, 50, 10) + rnorm(100, 0, 2)
m1 <- mean(x1)
se1 <- sqrt(var(x1))
x2 <- rnorm(100, 75, 15) + rnorm(100, 0, 5)
m2 <- mean(x2)
se2 <- sqrt(var(x2))

#fixed effect model
combine.est(x=c(m1, m2), x.se=c(se1, se2), hetero=FALSE)
#random effect model
combine.est(x=c(m1, m2), x.se=c(se1, se2), hetero=TRUE)

}
\references{
Cochrane, W. G. (1954) "The combination of estimates from different
experiments", Biometrics, 10, pages 101–129.
}
\seealso{
\link{test.hetero.est}
}
