\name{concordanceIndexC}
\alias{concordanceIndexC}
\title{
Concordance index C
}
\description{
Concordance index C
}
\arguments{
\item{msurv}{...}
\item{ustrat}{...}
\item{x2}{...}
\item{cl2}{...}
\item{st}{...}
\item{se}{...}
\item{weights}{...}
\item{strat}{...}
\item{N}{...}
\item{outx}{...}
\item{ch}{...}
\item{dh}{...}
\item{uh}{...}
\item{rph}{...}
\item{lenS}{...}
\item{lenU}{...}
}
\author{
Catharina Colsen, Zhaleh Safikhani, Benjamin Haibe-Kains
}
\keyword{ methods }
