% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D.index.R
\name{D.index}
\alias{D.index}
\title{Function to compute the D index}
\usage{
D.index(x, surv.time, surv.event, weights, strat, alpha = 0.05,
  method.test = c("logrank", "likelihood.ratio", "wald"), na.rm = FALSE, ...)
}
\arguments{
\item{x}{a vector of risk predictions.}

\item{surv.time}{a vector of event times.}

\item{surv.event}{a vector of event occurrence indicators.}

\item{weights}{weight of each sample.}

\item{strat}{stratification indicator.}

\item{alpha}{apha level to compute confidence interval.}

\item{method.test}{Statistical test to use in order to compute the p-values
related to a D. index, see \link[survival:summary.coxph]{survival::summary.coxph} for more details.}

\item{na.rm}{\code{TRUE} if missing values should be removed.}

\item{...}{additional parameters to be passed to the \link[survival:coxph]{survival::coxph}
function.}
}
\value{
A list with items:
\itemize{
\item d.index: D index (exponentiated, aka hazard ratio).
\item coef: D index estimate (coefficient) fitted in the cox regression model.
\item se: standard error of the estimate.
\item lower: lower bound for the confidence interval.
\item upper: upper bound for the confidence interval.
\item p.value: p-value for the statistical test if the estimate if different from 0.5.
\item n: number of samples used for the estimation.
\item coxm: \link[survival:coxph.object]{survival::coxph.object} fitted on the survival data and z (see below).
\item data: list of data used to compute the index (x, z, surv.time and
surv.event). The item z contains the scaled rankits which are the expected
standard Normal order statistics scaled by kappa.
}
}
\description{
Function to compute the D index for a risk prediction, i.e. an estimate of
the log hazard ratio comparing two equal-sized prognostic groups. This is
a natural measure of separation between two independent survival
distributions under the proportional hazards assumption.
}
\details{
The D index is computed using the Cox model fitted on the scaled rankits of
the risk scores instead of the risk scores themselves. The scaled rankits
are the expected standard Normal order statistics scaled by
kappa = sqrt(8/pi). See (Royston and Sauerbrei, 2004) for details.
Note that the value D reported in (Royston and Sauerbrei, 2004) is given.
}
\references{
Royston, P. and Sauerbrei, W. (2004) "A new measure of prognostic separation
in survival data", Statistics in Medicine, 23, pages 723–748.
}
\seealso{
\link[survival:coxph]{survival::coxph}, \link[survival:coxph.object]{survival::coxph.object}, \link[SuppDists:NormalScore]{SuppDists::normOrder}
}
