% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.all.parents.R
\name{.extract.all.parents}
\alias{.extract.all.parents}
\title{.extract.all.parents}
\usage{
.extract.all.parents(data, res.main, maxparents, predn)
}
\arguments{
\item{res.main}{output of regrnet.ensemble}

\item{maxparents}{maxparents parameter of the netinf method}

\item{predn}{list of target variables for which ensemble method was run.}
}
\description{
Function taking the output of the regrnet.ensemble method and returns a matrix
containing one equivalent model in each column. The target variable is in the first row.
}
\keyword{internal}
