\name{get_concordanceIndex_onevariable}
\alias{get_concordanceIndex_onevariable}
\title{
Get concordance index
}
\description{
Get the concordance index
}
\arguments{
\item{Rmsurv}{...}
\item{Rustrat}{...}
\item{Rx2}{...}
\item{Rcl2}{...}
\item{Rst}{...}
\item{Rse}{...}
\item{Rweights}{...}
\item{Rstrat}{...}
\item{RN}{...}
\item{Routx}{...}
\item{RlenS}{...}
\item{RlenU}{...}
}
\author{
Catharina Colsen
}
\keyword{ methods }
