% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsurv2.R
\name{getsurv2}
\alias{getsurv2}
\title{Function to retrieve the survival probabilities at a specific point
in time}
\usage{
getsurv2(sf, time, which.est = c("point", "lower", "upper"))
}
\arguments{
\item{sf}{survfit object}

\item{time}{time at which the survival probabilities must be retrieved}

\item{which.est}{which estimation to be returned? point for the point
estimate, lower for the lower bound and upper for the upper bound}
}
\value{
vector of survival probabilities
}
\description{
The function retrieves the survival probabilities from a survfit object, for
a specific point in time.
}
\details{
The survival probabilities are estimated through the \link[survival:survfit]{survival::survfit}
function.
}
\examples{
set.seed(12345)
age <- rnorm(30, 50, 10)
stime <- rexp(30)
cens <- runif(30,.5,2)
sevent <- as.numeric(stime <= cens)
stime <- pmin(stime, cens)
sf <- survfit(Surv(stime, sevent) ~ 1)
getsurv2(sf, time=1)

}
\seealso{
\link[survival:survfit]{survival::survfit}
}
