% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmr.cindex.R
\name{mrmr.cindex}
\alias{mrmr.cindex}
\title{Function to compute the concordance index for survival or binary class
prediction}
\usage{
mrmr.cindex(x, surv.time, surv.event, cl, weights, comppairs=10, strat,
alpha = 0.05, outx = TRUE, method = c("conservative", "noether", "nam"),
alternative = c("two.sided", "less", "greater"), na.rm = FALSE)
}
\arguments{
\item{x}{a vector of risk predictions.}

\item{surv.time}{a vector of event times.}

\item{surv.event}{a vector of event occurence indicators.}

\item{cl}{a vector of binary class indicators.}

\item{weights}{weight of each sample.}

\item{comppairs}{threshold for compairable patients.}

\item{strat}{stratification indicator.}

\item{alpha}{apha level to compute confidence interval.}

\item{outx}{set to \code{TRUE} to not count pairs of observations tied on x as a
relevant pair. This results in a Goodman-Kruskal gamma type rank
correlation.}

\item{method}{can take the value conservative, noether or name (see paper
Pencina et al. for details).}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" (concordance index is
greater than 0.5) or "less" (concordance index is less than 0.5). You can
specify just the initial letter.}

\item{na.rm}{TRUE if missing values should be removed.}
}
\value{
A mRMR ranking
}
\description{
Function to compute the minimum redundancy - maximum relevance (mRMR)
ranking for a risk prediction or a binary classification task based on the
concordance index. The mRMR feature selection has been adapted to use the
concordance index to estimate the correlation between a variable and the
output (binary or survival) data.
}
\section{Note}{

The "direction" of the concordance index (< 0.5 or > 0.5) is the opposite
than the \link[Hmisc:rcorr.cens]{Hmisc::rcorr.cens} function in the Hmisc package. So you can
easily get the same results than \link[Hmisc:rcorr.cens]{Hmisc::rcorr.cens} by changing the sign
of x.
}

\examples{
set.seed(12345)
age <- rnorm(100, 50, 10)
sex <- sample(0:1, 100, replace=TRUE)
stime <- rexp(100)
cens   <- runif(100,.5,2)
sevent  <- as.numeric(stime <= cens)
stime <- pmin(stime, cens)
strat <- sample(1:3, 100, replace=TRUE)
weight <- runif(100, min=0, max=1)
comppairs <- 10
xx <- data.frame("age"=age, "sex"=sex)
cat("survival prediction:\n")
mrmr.cindex(x=xx, surv.time=stime, surv.event=sevent, strat=strat, weights=weight,
method="noether", comppairs=comppairs)

}
\references{
Harrel Jr, F. E. and Lee, K. L. and Mark, D. B. (1996) "Tutorial in
biostatistics: multivariable prognostic models: issues in developing
models, evaluating assumptions and adequacy, and measuring and reducing
error", Statistics in Medicine, 15, pages 361–387.

Pencina, M. J. and D'Agostino, R. B. (2004) "Overall C as a measure of
discrimination in survival analysis: model specific population value and
confidence interval estimation", Statistics in Medicine, 23, pages
2109–2123, 2004.
}
\seealso{
\link[Hmisc:rcorr.cens]{Hmisc::rcorr.cens}, \link[CPE:phcpe]{CPE::phcpe}, \link[clinfun:coxphCPE]{clinfun::coxphCPE}
}
