\name{mrmr_cIndex_ensemble_remove}
\alias{mrmr_cIndex_ensemble_remove}
\title{
mRMR cIndex Ensemble Remove
}
\description{
mRMR cIndex Ensemble Remove
}
\arguments{
\item{Rdata}{...}
\item{Rnamat}{...}
\item{Rmaxparents}{...}
\item{Rnvar}{...}
\item{Rnsample}{...}
\item{Rpredn}{...}
\item{Rnpredn}{...}
\item{Rrep_boot}{...}
\item{Rmaxnsol}{...}
\item{Rthreshold}{...}
\item{Rmsurv}{...}
\item{Rustrat}{...}
\item{Rcl2}{...}
\item{Rst}{...}
\item{Rse}{...}
\item{Rweights}{...}
\item{Rstrat}{...}
\item{RN}{...}
\item{Routx}{...}
\item{RlenS}{...}
\item{RlenU}{...}
}
\author{
Heather Colsen
}
\keyword{ methods }
