% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no.at.risk.R
\name{no.at.risk}
\alias{no.at.risk}
\title{Function to compute the number of individuals at risk}
\usage{
no.at.risk(formula.s, data.s, sub.s = "all", t.step, t.end)
}
\arguments{
\item{formula.s}{formula composed of a Surv object and a strata variable
(i.e. stratification).}

\item{data.s}{data frame composed of the variables used in the formula.}

\item{sub.s}{vector of booleans specifying if only a subset of the data
should be considered.}

\item{t.step}{time step at which the number of individuals at risk is
computed.}

\item{t.end}{maximum time to be considered.}
}
\value{
number of individuals at risk at each time step specified in t.step up to
t.end.
}
\description{
Function to compute the number of individuals at risk at certain time points,
as used in the Kaplan-Meier estimator for instance, depending on
stratification.
}
\details{
The original version of this function was kindly provided by Dr Christos
Hatzis (January, 17th 2006).
}
\examples{
set.seed(12345)
stime <- rexp(100)
cens   <- runif(100,.5,2)
sevent  <- as.numeric(stime <= cens)
stime <- pmin(stime, cens)
strat <- sample(1:3, 100, replace=TRUE)
dd <- data.frame("surv.time"=stime, "surv.event"=sevent, "strat"=strat)
no.at.risk(formula.s=Surv(surv.time,surv.event) ~ strat, data.s=dd,
  sub.s="all", t.step=0.05, t.end=1)

}
\seealso{
\link[survival:survfit]{survival::survfit}, \link{km.coxph.plot}
}
