% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbrier.score2proba.R
\name{sbrier.score2proba}
\alias{sbrier.score2proba}
\title{Function to compute the BSCs from a risk score, for all the times of
event occurrence}
\usage{
sbrier.score2proba(data.tr, data.ts, method = c("cox", "prodlim"))
}
\arguments{
\item{data.tr}{the data frame for the training set. This data frame must
contain three columns for the times, the event occurrence and the risk
score. These columns are called "time", "event" and "score" respectively.}

\item{data.ts}{the data frame for the test set. This data frame must contain
three columns for the times, the event occurrence and the risk score.
These columns are called "time", "event" and "score" respectively.}

\item{method}{method for survival probabilities estimation using either a
Cox model or the product-limit estimator}
}
\value{
A list containing items:
\itemize{
\item time: vector of points in time
\item bsc: vector of Brier scores (BSC) at ome points in time
\item bsc.integrated: value of the integrated Brier score (IBSC)
}
}
\description{
The function computes all the Brier scores (BSC) and the corresponding
integrated Briser score (IBSC) from a risk score, for all the times of
event occurrence. The risk score is first transformed in survival
probabilities using either a Cox model or the product-limit estimator.
}
\examples{
set.seed(12345)
age <- rnorm(30, 50, 10)
stime <- rexp(30)
cens <- runif(30,.5,2)
sevent <- as.numeric(stime <= cens)
stime <- pmin(stime, cens)
dd <- data.frame("time"=stime, "event"=sevent, "score"=age)

#Cox's model
sbrier.score2proba(data.tr=dd, data.ts=dd, method="cox")
#product-limit estimator
sbrier.score2proba(data.tr=dd, data.ts=dd, method="prodlim")

}
\references{
Brier, G. W. (1950) "Verification of forecasts expressed in terms of
probabilities", Monthly Weather Review, 78, pages 1–3.

Graf, E. and Schmoor, C. and Sauerbrei, W. and Schumacher, M. (1999)
"Assessment and comparison of prognostic classification schemes for
survival data ", Statistics in Medicine, 18, pages 2529–2545.

Cox, D. R. (1972) "Regression Models and Life Tables", Journal of the Royal
Statistical Society Series B, 34, pages 187–220.

Andersen, P. K. and Borgan, O. and Gill, R. D. and Keiding, N. (1993)
"Statistical Models Based on Counting Processes", Springer.
}
\seealso{
\link[ipred:sbrier]{ipred::sbrier}, \link[survival:coxph]{survival::coxph}, \link[prodlim:prodlim]{prodlim::prodlim}
}
