\name{survcomp-package}
\alias{survcomp-package}
\alias{survcomp}
\docType{package}
\title{
Performance Assessment and Comparison for Survival Analysis
}
\description{
Functions to perform the performance assessment and comparison of risk prediction (survival) models.
}
\details{
\tabular{ll}{
Package: \tab survcomp\cr
Type: \tab Package\cr
Version: \tab \Sexpr{packageDescription("survcomp")$Version}\cr
Date: \tab \Sexpr{packageDescription("survcomp")$Date}\cr
License: \tab Artistic-2.0\cr
}
}

\author{
\bold{Benjamin Haibe-Kains}

- Bioinformatics and Computational Genomics Laboratory, Princess Margaret Cancer Center, Uni- versity Health Network, Toronto, Ontario, Canada

\url{http://www.pmgenomics.ca/bhklab/}

Former labs:

- Computational Biology and Functional Genomics Laboratory, Dana-Farber Cancer Institute, Boston, MA, USA

\url{http://compbio.dfci.harvard.edu/}

- Center for Cancer Computational Biology, Dana-Farber Cancer Institute, Boston, MA, USA

\url{http://cccb.dfci.harvard.edu/index.html}

- Machine Learning Group (MLG), Universite Libre de Bruxelles, Bruxelles, Belgium 

\url{http://www.ulb.ac.be/di/mlg/}

- Breast Cancer Translational Research Laboratory (BCTRL), Institut Jules Bordet, Bruxelles, Belgium 

\url{http://www.bordet.be/en/services/medical/array/practical.htm}

\bold{Markus Schroeder}

- UCD School of Biomolecular and Biomedical Science, Conway Institute, University College Dublin, Belfield, Dublin, Ireland

\url{http://bioinfo-casl.ucd.ie/cib/people/students/markusschroder}

\bold{Maintainer}:

Benjamin Haibe-Kains

\email{benjamin.haibe.kains@utoronto.ca}

Markus Schroeder

\email{markus.schroeder@ucdconnect.ie}
}
%\references{}
\keyword{ survival }
\keyword{ htest }
\seealso{
\code{survival}, \code{Hmisc}, \code{Design}, \code{prodlim}, \code{survivalROC}, \code{ipred}, \code{bootstrap}, \code{clinfun}
}
%\examples{}
