% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor.time.R
\name{censor.time}
\alias{censor.time}
\title{Function to artificially censor survival data}
\usage{
censor.time(surv.time, surv.event, time.cens = 0)
}
\arguments{
\item{surv.time}{vector of times to event occurrence}

\item{surv.event}{vector of indicators for event occurrence}

\item{time.cens}{point in time at which the survival data must be censored}
}
\value{
A list with the items:
\itemize{
\item surv.time.cens: vector of censored times to event occurrence
\item srv.event.cens: vector of censored indicators for event occurrence
}
}
\description{
The function censors the survival data at a specific point in time. This is
is useful if you used datasets having different follow-up periods.
}
\examples{
set.seed(12345)
stime <- rexp(30)
cens <- runif(30,0.5,2)
sevent <- as.numeric(stime <= cens)
stime <- pmin(stime, cens)
censor.time(surv.time=stime, surv.event=sevent, time.cens=1)

}
