% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.test.R
\name{combine.test}
\alias{combine.test}
\title{Function to combine probabilities}
\usage{
combine.test(p, weight, method = c("fisher", "z.transform", "logit"),
  hetero = FALSE, na.rm = FALSE)
}
\arguments{
\item{p}{vector of p-values}

\item{weight}{vector of weights (e.g. sample size of each study)}

\item{hetero}{fisher for the Fisher's combined probability test, z.transform
for the Z-transformed test, logit for the weighted Z-method}

\item{na.rm}{TRUE if the missing values should be removed from the data,
FALSE otherwise}
}
\value{
\code{\link{numeric}} p-value
}
\description{
The function combines several p-value estimated from the same null hypothesis
in different studies involving independent data.
}
\details{
The p-values must be one-sided and computed from the same null
hypothesis.
}
\examples{
p <- c(0.01, 0.13, 0.07, 0.2)
w <- c(100, 50, 200, 30)

#with equal weights
combine.test(p=p, method="z.transform")
#with p-values weighted by the sample size of the studies
combine.test(p=p, weight=w, method="z.transform")

}
\references{
Whitlock, M. C. (2005) "Combining probability from independent tests: the
weighted Z-method is superior to Fisher's approach", J. Evol. Biol., 18,
pages 1368–1373.
}
\seealso{
\link{test.hetero.test}
}
