% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.ratio.R
\name{hazard.ratio}
\alias{hazard.ratio}
\title{Function to estimate the hazard ratio through Cox regression}
\usage{
hazard.ratio(x, surv.time, surv.event, weights, strat, alpha = 0.05,
method.test = c("logrank", "likelihood.ratio", "wald"), na.rm = FALSE, ...)
}
\arguments{
\item{x}{a vector of risk predictions.}

\item{surv.time}{a vector of event times.}

\item{surv.event}{a vector of event occurrence indicators.}

\item{weights}{weight of each sample.}

\item{strat}{stratification indicator.}

\item{alpha}{alpha level to compute confidence interval.}

\item{method.test}{Statistical test to use in order to compute the p-values
related to a D.index, see \link[survival:summary.coxph]{survival::summary.coxph} for more details.}

\item{na.rm}{\code{TRUE} if missing values should be removed.}

\item{...}{additional parameters to be passed to the \link[survival:coxph]{survival::coxph}
function.}
}
\value{
A list with items:
\itemize{
\item hazard.ratio: hazard ratio estimate.
\item coef: coefficient (beta) estimated in the cox regression model.
\item se: standard error of the coefficient (beta) estimate.
\item lower: upper bound for the confidence interval.
\item upper: upper bound for the confidence interval.
\item p.value: p-value computed using the likelihood ratio test whether the hazard ratio is different from 1.
\item n: number of samples used for the estimation.
\item data: list of data used to compute the hazard ratio (x, surv.time and
surv.event).
}
}
\description{
Function to compute the hazard ratio for a risk prediction.
}
\details{
The hazard ratio is computed using the Cox model.
}
\examples{
set.seed(12345)
age <- rnorm(100, 50, 10)
stime <- rexp(100)
cens   <- runif(100,.5,2)
sevent  <- as.numeric(stime <= cens)
stime <- pmin(stime, cens)
strat <- sample(1:3, 100, replace=TRUE)
weight <- runif(100, min=0, max=1)
hazard.ratio(x=age, surv.time=stime, surv.event=sevent, weights=weight,
  strat=strat)

}
\references{
Cox, D. R. (1972) "Regression Models and Life Tables", Journal of the Royal
Statistical Society Series B, 34, pages 187–220.
}
\seealso{
\link[survival:coxph]{survival::coxph}, \link[survival:coxph.object]{survival::coxph.object}
}
