% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr.comp.R
\name{hr.comp}
\alias{hr.comp}
\title{Function to statistically compare two hazard ratios}
\usage{
hr.comp(hr1, hr2)
}
\arguments{
\item{hr1}{first hazard ratio.}

\item{hr2}{second hazard ratio.}
}
\value{
A list with items:
\itemize{
\item p.value: p-value from the Student t test for the comparison beta1 > beta2
(equivalently hr1 > hr2)
\item hr1: value of the first hazard ratio
\item hr2: value of the second hazard ratio
}
}
\description{
This function compares two hazard ratios from their betas and standard
errors as computed by a Cox model for instance. The statistical test is
a Student t test for dependent samples. The two hazard ratios must be
computed from the same survival data.
}
\details{
The two hazard ratios must be computed from the same samples (and
corresponding survival data). The function uses a Student t test for
dependent samples.
}
\examples{
set.seed(12345)
age <- as.numeric(rnorm(100, 50, 10) >= 50)
size <- as.numeric(rexp(100,1) > 1)
stime <- rexp(100)
cens <- runif(100,.5,2)
sevent <- as.numeric(stime <= cens)
stime <- pmin(stime, cens)
hr1 <- hazard.ratio(x=age, surv.time=stime, surv.event=sevent)
hr2 <- hazard.ratio(x=size, surv.time=stime, surv.event=sevent)
hr.comp(hr1=hr1, hr2=hr2)

}
\references{
Student (1908). "The Probable Error of a Mean", Biometrika, 6, 1, pages 1–25.
}
\seealso{
\link[survival:coxph]{survival::coxph}, \link[stats:t.test]{stats::t.test}
}
