% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iauc.comp.R
\name{iauc.comp}
\alias{iauc.comp}
\title{Function to compare two IAUCs through time-dependent ROC curves}
\usage{
iauc.comp(auc1, auc2, time)
}
\arguments{
\item{auc1}{vector of AUCs computed from the first time-dependent ROC curves
for somepoints in time}

\item{auc2}{c2vector  of  AUCs  computed  from  the  second  time-dependent
ROC  curves  forsome points in time}

\item{time}{vector of points in time for which the AUCs are computed}
}
\value{
A list with items:
\itemize{
\item p.value: p-value from the Wilcoxon rank sum test for the comparison
iauc1 > iauc2
\item iauc1: value of the IAUC for the first set of time-dependent ROC curves
\item iauc2: value of the IAUC for the second set of time-dependent ROC curves
}
}
\description{
This function compares two integrated areas under the curves (IAUC) through
the results of time-dependent ROC curves at some points in time. The
statistical test is a Wilcoxon rank sum test fordependent samples.
}
\details{
The two vectors of AUCs must be computed from the same samples (and
corresponding survivaldata) and for the same points in time. The function
uses a Wilcoxon rank sum test for dependent samples.
}
\examples{
set.seed(12345)
age <- rnorm(30, 50, 10)
size <- rexp(30,1)
stime <- rexp(30)
cens <- runif(30,.5,2)
sevent <- as.numeric(stime <= cens)
stime <- pmin(stime, cens)
##time-dependent ROC curves
tt <- unique(sort(stime[sevent == 1]))
##size
mytdroc1 <- NULL
for(i in 1:length(tt)) {
	  rr <- tdrocc(x=size, surv.time=stime, surv.event=sevent, time=tt[i], na.rm=TRUE, verbose=FALSE)
	  mytdroc1 <- c(mytdroc1, list(rr))
}
auc1 <- unlist(lapply(mytdroc1, function(x) { return(x$AUC) }))
##age
mytdroc2 <- NULL
for (i in 1:length(tt)) {
	  rr <- tdrocc(x=age, surv.time=stime, surv.event=sevent, time=tt[i], na.rm=TRUE, verbose=FALSE)
   mytdroc2 <- c(mytdroc2, list(rr))
}
auc2 <- unlist(lapply(mytdroc2, function(x) { return(x$AUC) }))
iauc.comp(auc1=auc1, auc2=auc2, time=tt)

}
\references{
Wilcoxon, F. (1945) "Individual comparisons by ranking methods", Biometrics
Bulletin,1, pages80–83

Haibe-Kains, B. and Desmedt, C. and Sotiriou, C. and Bontempi, G. (2008) "A
comparative studyof survival models for breast cancer prognostication
based on microarray data:  does a single genebeat them all?",
Bioinformatics, 24, 19, pages 2200–2208.
}
\seealso{
\link{tdrocc}, \link[stats:wilcox.test]{stats::wilcox.test}
}
