\name{survcomp-package}
\alias{survcomp-package}
\alias{survcomp}
\docType{package}
\title{
Performance Assessment and Comparison for Survival Analysis
}
\description{
Functions to perform the performance assessment and comparison of risk prediction (survival) models.
}
\details{
\tabular{ll}{
Package: \tab survcomp\cr
Type: \tab Package\cr
Version: \tab 1.3.4\cr
Date: \tab 2011-06-24\cr
License: \tab Artistic-2.0\cr
}
}

\author{
\bold{Benjamin Haibe-Kains}

- Computational Biology and Functional Genomics Laboratory, Dana-Farber Cancer Institute, Boston, MA, USA

\url{http://compbio.dfci.harvard.edu/}

- Center for Cancer Computational Biology, Dana-Farber Cancer Institute, Boston, MA, USA

\url{http://cccb.dfci.harvard.edu/index.html}

Former labs:

- Machine Learning Group (MLG), Universite Libre de Bruxelles, Bruxelles, Belgium 

\url{http://www.ulb.ac.be/di/mlg/}

- Breast Cancer Translational Research Laboratory (BCTRL), Institut Jules Bordet, Bruxelles, Belgium 

\url{http://www.bordet.be/en/services/medical/array/practical.htm}

\bold{Markus Schroeder}

- Computational Biology and Functional Genomics Laboratory, Dana-Farber Cancer Institute, Boston, MA, USA

\url{http://compbio.dfci.harvard.edu/}

- Computational Genomics, Center for Biotechnology (CeBiTec), Bielefeld University, Germany

\url{http://www.cebitec.uni-bielefeld.de/cebitec/computational-genomics/home.html}

\bold{Maintainer}:

Benjamin Haibe-Kains

\email{bhaibeka@jimmy.harvard.edu}

\email{bhaibeka@ulb.ac.be}

Markus Schroeder

\email{mschroed@jimmy.harvard.edu}
}
%\references{}
\keyword{ survival }
\keyword{ htest }
\seealso{
\code{survival}, \code{Hmisc}, \code{Design}, \code{prodlim}, \code{survivalROC}, \code{ipred}, \code{bootstrap}, \code{CPE}, \code{clinfun}
}
%\examples{}
