\name{survcomp-package}
\alias{survcomp-package}
\alias{survcomp}
\docType{package}
\title{
Performance Assessment and Comparison for Survival Analysis
}
\description{
Functions to perform the performance assessment and comparison of risk prediction (survival) models.
}
\details{
\tabular{ll}{
Package: \tab survcomp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-02-08\cr
License: \tab GPL-2\cr
}
}

\author{
Benjamin Haibe-Kains

- Machine Learning Group (MLG), Universite Libre de Bruxelles, Bruxelles, Belgium \url{http://www.ulb.ac.be/di/mlg/}

- Functional Genomics Unit (FGU), Institut Jules Bordet, Bruxelles, Belgium \url{http://www.bordet.be/en/services/medical/array/practical.htm}

Maintainer: Benjamin Haibe-Kains <\email{bhaibeka@ulb.ac.be}>
}
%\references{}
\keyword{ survival }
\keyword{ htest }
\seealso{
\code{survival}, \code{Hmisc}, \code{Design}, \code{prodlim}, \code{survivalROC}, \code{ipred}
}
%\examples{}
