% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdrocc.R
\name{tdrocc}
\alias{tdrocc}
\title{Function to compute time-dependent ROC curves}
\usage{
tdrocc(
  x,
  surv.time,
  surv.event,
  surv.entry = NULL,
  time,
  cutpts = NA,
  na.rm = FALSE,
  verbose = FALSE,
  span = 0,
  lambda = 0,
  ...
)
}
\arguments{
\item{surv.time}{vector of times to event occurrence.}

\item{surv.event}{vector of event occurrence indicators.}

\item{time}{time point for sensitivity and specificity estimations.}

\item{na.rm}{\code{TRUE} if the missing values should be removed from the data,
\code{FALSE} otherwise.}

\item{span}{Span for the NNE. Default value is 0.}

\item{...}{additional arguments to be passed to the
\link[survivalROC:survivalROC]{survivalROC::survivalROC} function.}

\item{cl}{vector of binary classes.}

\item{sampling}{jackknife procedure to estimate the standard error of
sensitivity and specificity estimations.}
}
\value{
A list with items:
\itemize{
\item sens: sensitivity estimate
\item sens.se: standard error for sensitivity estimate
\item spec: specificity estimate
\item spec.se: standard error for specificity estimate
}
}
\description{
The function is a wrapper for the survivalROC.C function in order to compute
sensitivity and specificity for a binary classification of survival data.
}
\details{
Only NNE method is used to estimate sensitivity and specificity
(see \link[survivalROC:survivalROC.C]{survivalROC::survivalROC.C}). The standard error for sensitivity and
specificity is estimated through jackknife procedure
(see \link[bootstrap:jackknife]{bootstrap::jackknife}).
}
\examples{
set.seed(12345)
gender <- sample(c(0,1), 100, replace=TRUE)
stime <- rexp(100)
cens <- runif(100,.5,2)
sevent <- as.numeric(stime <= cens)
stime <- pmin(stime, cens)
mysenspec <- td.sens.spec(cl=gender, surv.time=stime, surv.event=sevent,
  time=1, span=0, na.rm=FALSE)

}
\references{
Heagerty, P. J. and Lumley, T. L. and Pepe, M. S. (2000) "Time-Dependent ROC
Curves for Censored Survival Data and a Diagnostic Marker", Biometrics,
56, pages 337–344.

Efron, B. and Tibshirani, R. (1986). "The Bootstrap Method for standard
errors, confidence intervals, and other measures of statistical accuracy",
Statistical Science, 1 (1), pages 1–35.
}
\seealso{
\link[survivalROC:survivalROC]{survivalROC::survivalROC}
}
