% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.hetero.est.R
\name{test.hetero.est}
\alias{test.hetero.est}
\title{Function to test the heterogeneity of set of probabilities}
\usage{
test.hetero.est(x, x.se, na.rm = FALSE)
}
\arguments{
\item{x}{vector of estimates}

\item{x.se}{vector of standard errors of the corresponding estimates}

\item{na.rm}{\code{TRUE} if the missing values should be removed from the data,
\code{FALSE} otherwise}
}
\value{
A list with items:
\itemize{
\item Q: Q statistic
\item p.value: p-value of the heterogeneity test
}
}
\description{
The heterogeneity test is known to be very conservative. Consider a
p-value < 0.1 as significant.
}
\details{
The heterogeneity test is known to be very conservative. Consider a
p-value < 0.1 as significant.
}
\examples{
set.seed(12345)
x1 <- rnorm(100, 50, 10) + rnorm(100, 0, 2)
m1 <- mean(x1)
se1 <- sqrt(var(x1))
x2 <- rnorm(100, 75, 15) + rnorm(100, 0, 5)
m2 <- mean(x2)
se2 <- sqrt(var(x2))

test.hetero.est(x=c(m1, m2), x.se=c(se1, se2))

}
\references{
Cochrane, W. G. (1954) "The combination of estimates from different
experiments", Biometrics, 10, pages 101–129.
}
\seealso{
\link{combine.est}
}
