% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.hetero.test.R
\name{test.hetero.test}
\alias{test.hetero.test}
\title{Function to test the heterogeneity of set of probabilities}
\usage{
test.hetero.test(p, weight, na.rm = FALSE)
}
\arguments{
\item{p}{vector of p-values}

\item{weight}{vector of weights (e.g. sample size of each study)}

\item{na.rm}{\code{TRUE} if the missing values should be removed from the data,
\code{FALSE} otherwise}
}
\value{
A list with items:
\itemize{
\item Q: Q statistic
\item p.value: p-value of the heterogeneity test
}
}
\description{
The function tests whether a set of p-values are heterogeneous.
}
\details{
The p-values should be one-sided and computed from the same null hypothesis.
}
\examples{
p <- c(0.01, 0.13, 0.07, 0.2)
w <- c(100, 50, 200, 30)

#with equal weights
test.hetero.test(p=p)
#with p-values weighted by the sample size of the studies
test.hetero.test(p=p, weight=w)

}
\references{
Cochrane, W. G. (1954) "The combination of estimates from different
experiments", Biometrics, 10, pages 101–129.

Whitlock, M. C. (2005) "Combining probability from independent tests: the
weighted Z-method is superior to Fisher's approach", J. Evol. Biol., 18,
pages 1368–1373.
}
\seealso{
\link{combie.est}
}
